tmux: make default-command/default-shell generic
We likely don't need to set default-shell, but do it anyway. We set
default-command to be a bare $SHELL so that we use an interactive, but
not a login shell. If we don't set this, we'd get a login shell.

We prefer a non-login shell so that we can put tmux auto-connect in
the shell's login config and have it bypassed simply with new tmux
windows.