X/i3/polybar: Stop fiddling with dpi settings
Let the defaults roll and just ensure that font sizes for polybar and
i3 make sense.

Everything else we run picks things up and does sensible configuration
with them, so it looks and works just fine.

This solves the bug of needing to restart i3 after first load to have
things sized correctly.

Signed-off-by: Ben Walton <bdwalton@gmail.com>