emacs/installer: use package apheleia so we always format code on save
This means we don't need the specific go-mode hook any more because it
is superceded with something more globally useful.

This package wants clang-format for c/cc files, so ensure that's
always installed as well. (There might be other tools it needs, but
we'll add them as we identify them.)