emacs: change the way we handle TERM and COLORTERM
Because TERM leaks through to a sudo call by default, we don't want to
globally export TERM=xterm-24bit as that will break some things we
care about. Instead, run emacs via a small wrapper script.

Because we now have a wrapper script, we can also move COLORTERM to
that script because emacs is the only thing using it.

Signed-off-by: Ben Walton <bdwalton@gmail.com>