emacs/i3: Integrate emacs systemd startup/shutdown with i3
Now that we have a systemd service, we can make i3 start the emacs
daemon for us at every login and shut it down when we log out. This
requires a wrapper script for startup to ensure we clear any previous
failures when starting it. We see those right now and will investigate
why shutdown isn't clean later. To shut it down, we have i3 spin up a
program that listens for an IPC message that it's terminating and then
call systemd to shutdown the user service.

This requires the py3 i3ipc package.

Signed-off-by: Ben Walton <bdwalton@gmail.com>