installer/systemd: Deliver systemd service files more cleanly.
Systemd wasn't liking the way I was using symlinks for the service
file delivery and that was making things brittle. Instead of
symlinking my systemd service files into dotfiles, hardlink
them.

Update the installer script and make it generic as we'll be delivering
more than just an emacs service shortly.