ssh: Stop using keychain, leverage systemd and new ssh features
We have an agent running, via systemd, now and a standard
SSH_AUTH_SOCK that should always be visible and available. There is
also a feature in ssh 7.2+ that allows us to store keys in an agent
after the first unlock.

With this in place, the next change will be to start leveraging the
new ssh feature. That will be nicer if we import our external
ssh_config git repo though, so it will be a followup change.

Signed-off-by: Ben Walton <bdwalton@gmail.com>