installer: No longer require shfmt.
This package isn't on Debian bullseye and it's been a pain to find a
useful backport. As we have the package on most machines where we want
it, stop forcing it to be installed to unbreak ./setup-env on older
machines.

Once the next Debian release is out and we're updated, we can bring it
back. In the meantime, emacs won't be able to format shell scripts in
some places, but that was already true, so not a regression.