installer/shell: Don't export BDW_CONFIG_TYPE from ~/.bdwconfig
If we don't export it here, the ~/.bdwconfig file becomes something we
can make systemd consume via an EnvironmentFile setting. This means
that we can then use PassEnvironment when needed too. Both emacs and
tmux can benefit from that.

Because we don't export it when it is sourced now, we ensure it is
exported after sourcing so it is available to child processes later.

Signed-off-by: Ben Walton <bdwalton@gmail.com>