systemd: Source ~/.bdwconfig for environment and use it.
By sourcing ~/.bdwconfig as an EnvironmentFile in service units, we
can then pass BDW_CONFIG_TYPE into those environments so they're able
to make logical decisions using that variable as needed. For tmux and
emacs, this is useful because we have environment-specific needs in
some systems where this config is deployed.