emacs: Provide a hook that runs on frame creation.
This is useful because some things like tool and menu bar settings may
only work in some modes, so run a bit of code each time a frame is
created so we can have a consistent ui (no tool or menu bars
anywhere). We also need to specify some font settings in a gui frame
that aren't relevant in console mode.