i3/emacs: Rollback the change that removes scatchpad show.
For some reason this change affects other parts of what I want to
happen when emacs is popped up. For now, revert the change while we
figure out how to make it work correctly. This means that monitor
changes will show scratchpad emacs upon change, if it had been
launched.