emacs: For each package, add an additional heading level
Delimit each package with an org heading and split the code chunks
into per-package blocks. This makes it nicer to capture any package
specific documentation appropriately.

Now, because we have 3 useful levels of indent, have org mode startup
with that 3rd level visible too. This gives us main blocks, groups of
packages and then individual packages which is a great birds eye view
of the config.