i3/polybar/autorandr: Fix the handling of multiple monitors
polybar: Have a primary and secondary bar config so we can start and
	 stop individual variants.
i3: Only run the polybar launcher at initial startup.
autorandr: Drop the predetect script, make the postswitch script
	   smarter. Trigger polybar process changes via i3-msg
launcher: Detect primary monitor and start the primary bar for it. All
	  other monitors start the new secondary bar.

With this setup, monitor switching is less janky and bars are started
and stopped cleanly.