installer: Ensure we get current nodejs
Add a script that sets up a debian repo config for nodejs 18.x. And
now, with this repo giving a modern deb, we also get npm from the
nodejs package. The older package in debian isn't compatible, so we
drop it.

Note that this is a difference in home and non-home environments. In
other environments, there is an updated nodejs and a separate but also
updated npm. So on a non-home machine, we'd need the local non-home
installer to grab npm. But we're doing this kind of dev on home
machines so it doesn't matter for now.

Signed-off-by: Ben Walton <bdwalton@gmail.com>