sway: Fork a copy of i3 config
Use the more canonical location for sway config so that we have 2
separate files that can be used. This will allow us to maintain both
environments in working order while we transition. This is important
because sway needs additional config for (eg) keyboard settings that
i3 used to delegate leverage X11 for.