sway: Improve screen idle/locking
Add a lock stanza to the swayidle command so that it handles session
locking by executing a command. Then, instead of exec'ing a custom
lock script on mod+shift+L, invoke loginctl to trigger a lock session
command. Overall this should be more natural within the context of a
systemd setup.