waybar/sway/inhibit-bridge: Use a new version of inhibit bridge.
This version has tray icon support to handle the manual toggle aspect
of things. There is some fragility around startup now, which I'll fix
later. Simply, the tray icon machinery only works if dbus has a
handler for org.kde.StatusNotifierWatcher. waybar provides this but
that means it has to be alive first.