emacs: Switch from company mode to corfu for completion.
Corfu is a more native integration (thus simpler) so switch to it as
it's more in line with the types of packages we're now preferring. At
present, corfu works for graphical frames, but doesn't show up the
completion menu selection in the terminal. This is a known
issue. We'll add corfu-terminal mode support next.