Sway: Better handle changing names for external displays
It seems that the external display name is unstable. I've hit this a
few times now and it's annoying. So, given the internal display is
fixed, we'll apply external and global settings to display '*' and
then override the positioning for the internal display. This _should_
be robust to name changing now.