waybar: Style visible but not focussed workspaces
Much like the scratchpad being visible in the bar when not active,
make visible but not focused workspaces stand out in the workspace
widget. Use the same color as for the scratchpad so there is
similarity for "not focused" things. This should make keyboard nav a
little bit nicer as we can know for sure which workspace is shown on
alternate monitors.