tmux: Adjust default window setup
Because we run this as a service, we were ending up with a shell in
window 0 that had incorrect SSH_AUTH_SOCK most of the time. "Fix" this
by ensuring we never have a default shell window running until first
login when we manually create one which will get a viable
SSH_AUTH_SOCK. This is still not pretty but it's slightly less ugly.