waybar: More color styling
Have volume and power indicator be the same as the clock (black
background, white text) but allow for color use when there is
something interesting about them. The power indicator will use
yellow (active) for charging and red for low still. The volume will be
red when muted. This makes them not catch the eye unless there is
something to see.